requirements.txt
kaggle
numpy
sklearn
ipykernel
flask
gunicorn
PyYAML
pandas
evidently
dill
matplotlib
pandas
openpyxl
-e .
